hydra: Move PostgreSQL server to ultron.
While taalo is the Hydra master, oltron is the frontend web server and
is going to be used as a DBMS server for use with MAM anyway (#14).

Another issue was that the DBMS and the evaluator plus local builds were
constantly fighting over I/O, which now shouldn't be the case anymore,
except of course if the same happens on ultron.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>