pkgs/mongooseim: Work around systemd-notify issue.
If systemd-notify is called too early, it could happen that the service
doesn't switch to the active state and stays in "activating" for all
eternity.

The issue has been reported upstream at:

http://comments.gmane.org/gmane.comp.sysutils.systemd.devel/18649

So we just wait for 10 seconds until we run systemd-notify to work
around this issue, but in the long run we probably want to use a NIF for
avoiding that ugly delay.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>