Add a module and test for a dynamic DNS service
This is very similar to what DynDNS, AnyDNS, afraid.org and others are
doing, but these services tend to turn into non-free services after a
while (well, of course afraid.org is still free) and we *already* run
our own DNS infrastructure, so why pay for something which you largely
already have?

My journey goes on to looking for FOSS implementations of such a such a
service, but they seem either rely too much on BIND, are too inflexible
and/or tailored to run on one nameserver only (which then would
IXFR/AXFR the changes to the slave nameservers) or are even running its
own nameserver (like minidyndns does).

Also I wanted to lower the attack surface as much as possible, so while
in general I didn't make any security precautions in the implementation
code, most of it is largely at a macro scale, for example letting the
processes communicate via an encrypted tunnel.

The implementation itself consists of a master and one or more slave
services. Communication between them is simply by broadcasting zone
updates from the master to all connected slaves. Once a slave loses the
connection and reconnects, it gets the same update packages but for all
available zones.

Both sides use plain TCP to communicate and no authentication
whatsoever, because the latter should be handled via other methods.

Of course, we could add support for TLS and whatnot, but that's
something to be done in future improvements.

The architecture of what I envision how it would be used in production
is like the following:

    _  __  _       .--------------.    .----------------.
  _(_)(__)(_)_ _-> | nameserver 1 | <- | dyndns slave 1 | <----.
 (_)internet(_)    `--------------'    `----------------'      |
   (_)(__)(_) `.   .--------------.    .----------------.      |
       |        `> | nameserver 2 | <- | dyndns slave 2 | <-.  |
      \|/          `--------------'    `----------------'   |  |
  .-----------.                          .-----------.      |  |
  | front-end |     .---------------.    | encrypted | -----'  |
  | webserver | --> | dyndns master | -> |           |         |
  | using TLS |     `---------------'    |  tunnel   | --------'
  `-----------'                          `-----------'

So as you see there are only the two nameservers and the front-end
webserver exposed to the internet, so we have lowered down the attack
surface to just the HTTP portion of the dyndns master process and I've
made sure that the implementation stays free of unnecessary stuff.

Now for example if you want to update the A resource record of a
particular FQDN, all you do is place a request with the correct user
credentials, the FQDN and IP (automatically determining the IP address
is not yet implewented). The front-end webserver basically just proxies
the HTTP server of the dyndns master server, so the request ending up
there is authenticated against a flat-file user/passwd database and if
everything is correct, the request is queued in a worker process and a
response is sent back to the client.

In summary, the worker process within the master just boils down to a
queue where every valid request generates a zone update in the queue
along with an increment of the current serial number (that's one of the
reason the master needs a state directory).

After broadcasting the zone updates to all of the currently connected
slaves (future slaves still get the current versions of *all* zones so
they stay up to date), the slaves themselves write the zone files to a
directory to pick up by the correspending DNS server software and
notifies the DNS server to reload/add the zone. All of this is done in a
queue as well, so we shouldn't run into race conditions, at least for
the queue itself.

Right now, the implementation has a few goofs we have to get rid off in
the long term:

 * Within the test (and reference) NixOS module implementation (using
   NSD) zone file updates are done using inotify on the zonefile itself
   to detect whether NSD has already picked up the zone. The reason for
   this is to prevent race conditions whenever there are multiple
   updates of the same zone in a sequence.
 * Zone files are update using *one* queue, so even if two
   different zones could be updated at once, we only update based on a
   queue among *all* zones. This of course only needs to be improved as
   soon as we decide to go large-scale.
 * In the NixOS module, user credentials are pasted verbatim into the
   master configuration file, which is not optimal because we may want
   to use an external file with permissions set accordingly for that.
 * The reference implementation is within the NixOS test, but it would
   make sense to factor it out into the main module or a different
   module wrapping the dyndns module so we can simplify our test code.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>