modules/dyndns: Move NSD integration into module
So far we only had the integration for NSD in the VM test but not yet in
the main module. Now there is a useNSD option which if it is true will
set up the slave so it communicates with a locally running NSD via
nsd-control.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
(cherry picked from commit 1c8c508f9fc3e9451635d5f5d2c5f4ce27ad355a)