dyndns/nsd: Provide a config file for nsd-control
So far we have been using a very ugly hack using systemctl to get the
current configuration file because it isn't exposed by the upstream
NixOS NSD module.

We only need a handful of the configuration values to keep nsd-control
happy and we already have them available via the NixOS module system, so
we can simply generate a trimmed-down nsd-config-only configuration
file.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>