modules/dyndns: Create special user for slave
So far only the master was running under the dyndns user and group,
because the slave had to run a zoneCommand, which usually doesn't have
the neccessary permissions to write to the nameserver directory.

From a security perspective however, this doesn't make sense and we
should really run it as a normal user and if some configuration wants to
use "root" for its user it can still define it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
(cherry picked from commit e98c5fd7cedd0829d6ca9f54d05f3db26cf96ef1)