modules/dyndns: Fix asynchronous serial numbers
We didn't have this in our tests, but if two or more slaves are
connected to the master, the serial number is increased for *each*
slave, because we just broadcasted information from the HTTP request to
the master workers that handle slave connections.

The latter did the increment of the serial numbers and also wrote the
new acid-state to disk.

Of course, we want consistent data across all nameservers, so we now
create the Zone information (and thus also the serial) within the
master's HTTP worker and propagate that through the broadcast channel
that is sending the stuff out to the slaves.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
(cherry picked from commit d6a2d1643e19a2ab4d4c5b749402e197ca290d2e)