tests/dyndns: Add a second DNS server
Since we have the ability to listen to multiple interfaces, it makes
sense to also test that functionality.

This is done by creating VLANs in conjunction to different eth*
interfaces which are dedicated to those VLANs, so we can make sure that
nothing gets "accidentally" sent to the wrong IP address.

It also improves our expectDNS subroutine to always make requests to
both nameservers, so as soon as zones get desynchronized our test will
(and should) fail.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>