pkgs: Update xmpppoke to latest upstream tip
We now fetch the sources using fetchFromBitbucket, because this only
fetches a tarball instead of cloning the repository. I've done this
mainly to avoid cloning submodules.

The new version now seems to require libunbound, so I've added it to the
requirements of xmpppoke so that it most closely resembles its squishy
file.

However, there still seem to be two issues:

 * DH params do not seem to get picked up, see log at:
   https://headcounter.org/hydra/log/dqng385dlnwd099m6i3v9nwrcw2jw1y2-vm-test-run-headcounter-vhost-headcounter.drv
 * The "Show PEM" button is dysfunctional in the front-end.

I also had to apply a small patch (server-handler.patch) which avoids an
exception that's being thrown because handler:close() is applied to a
handler that doesn't exist anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Issue: #23
Cc: @xnyhps