xmpp: Introduce new virtual host "torservers.net"
This replaces "jabber.headcounter.org", which has been obsolete since a
long time, because we're just using "headcounter.org" and there is no
user registered with the former vhost.

So the new torservers.net configuration needs to be handled a bit
differently, especially regarding admin rules/ACLs.

Soon we also need to provide an extra MUC host called
conference.torservers.net, which currently doesn't exist because we
forward the conference nodes for all the vhosts to headcounter.org.

The ACL for the admin address is currently marked with TODO along with
the others. Of course in the real deployment the "TODO"s are replaced
with the real admin addresses, because I don't want to disclose them
publicly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>