mongooseim: Update to upstream version 1.6.2
This is currently just a version bump without changes to the
dependencies (well, except adding usec to the direct dependencies of
mongooseim) but with updates of the patches for the main derivation and
the tests.

The tests will probably fail on a spectacular level right now, so we
need a bit more work on this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>