vhosts: Allow to set a global default device
This is especially cruicial for tests, because we want to assign *all*
vhosts to eth1 instead of the default (eth0). Previously we had a static
list for the domains to be tested, but they're no longer the only
domains that are used for XMPP, so we can now fix this by assigning a
default device.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>