vhosts: Use declarative config instead of scripted
If we someday want to switch over to use networkd instead of the
scripted networking setup, it's a good idea to actually use the NixOS
options for setting those addresses instead of adding our own commands
via a systemd unit.

So instead of the units we now simply use the
networking.interfaces.<name>.*.ip[46] options, which have been in NixOS
for quite a while.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>