mongooseim: Fix listener socket options regression
Regression introduced by the following upstream commit:

  esl/MongooseIM@19287d71a6c6af1fa5f728fd0f266a19f6c627de

The commit has removed Erlang R13B specific handling but forgot to pass
over the SockOpts list to gen_tcp:listen/2.

This was the reason our integration tests have failed, because in our
setup we bind to multiple different IP addresses. Leaving out the
SockOpts values causes MongooseIM to try binding on 0.0.0.0 for all
listeners, which of course is going to fail for our setup because the
address is already in use by another listener.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>