tests/mongooseim: Allow users to delete themselves
This seems to have changed in-between the MongooseIM versions and at
least the tests for mod_adhoc doesn't delete the registered users
afterwards.

Turns out that inband_registration_and_cancelation_allowed/2 in
mod_register.erl defaults to allowing noone if the access option isn't
set.

So in order for the tests to cleanup nicely after each other we need to
make sure we give access to anyone (the access rule "register" already
is there in the test config).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>