lib/erlang: Tread padding of IPv6 specially
In the previous commit we have replaced every empty string with a 0
digit in one of the IPv6 address nodes. However, this also means that
we'd have to take care that we only insert the real padding in the
middle of the list of nodes.

So if we have "::", the splitted output is ["" "" ""]. We can't do
padding for all three of these nodes, so we'd have to work only on the
part with init (tail x) but that would end up being way too complex.

What we're doing instead is just replacing every "::" by ":PAD:" in the
first step before the splitting and only pad when we encounter a "PAD",
so the splitted output for "::" will then end up being ["" "PAD" ""].

While this is easier to follow it also means that we have to insert a
few sanity checks, for example we need to check whether the address
already has something like :PAD: in it so the user can't cheat her way
into successful validation and we also make sure that :PAD: doesn't
exist twice.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>