modules: Add a stub Postfix service module
First of all, yes, there is already a Postfix module in NixOS, but I'm
not going to use it for several reasons:

  * Configuration is generated using string concatenation.
  * It needs to symlink files from the store to runtime directories.
  * It doesn't allow for changing master.cf very much (for example you
    can't remove services).
  * It relies on the Postfix SysV-style init file.
  * It uses the Postfix's master process for service supervision.

This implementation is still work-in-progress (only the NixOS side is
implemented the actual runtime doesn't work right now) but it improves
the situation in that it tears apart Postfix's own service manager.

We have systemd for that and on top of that, we have Nix, so we can
easily generate arbitrary unit files according to our configuration and
we can also check the validity of the configuration at evaluation time
so we don't get smacked in the face by some nasty runtime error.

Having every single Postfix service as a systemd socket-activated
service and NixOS on top of that has the following advantages:

  * We can generate automatic chroot environments without copying around
    stuff that could break things at runtime (like missing libraries).
  * Instead of just having a flag to run as root or unprivileged user,
    we now can run everything using an unprivileged user and use
    capabilities to allow certain operations with finer granularity.
  * Having an abstraction for users on NixOS allows us to even run
    all Postfix services with its own dedicated user.
  * Adding to that, we can further restrict this by using seccomp to
    allow only specific syscalls.
  * Configuration options are serialized using a Nix attribute set, so
    syntax errors in the config file are a thing of the past.
  * We have socket activation at early boot, even before filesystems are
    (re-)mounted.
  * Every generated unit can be fully customized using the
    systemd.{sockets,services} NixOS options, which allows for even more
    control like setting CPU or I/O priorities.
  * It's also very easy to implement multiple instances of Postfix, we
    just need to generate the units with a different prefix.
  * The attack vector is slightly decreased, because we already have
    systemd as a service manager and don't need another implementation
    with different bugs.
  * External Postfix services can now be implemented very easily.

Of course, our implementation has a few disadvantages:

  * The complexity on the Nix side is slightly bigger and thus it also
    increases the evaluation time.
  * Throwing away the Postfix master process means that we need rely on
    internals.
  * We need to patch Postfix more heavily and thus need to thoroughly
    test the stuff we're changing.
  * And of course, it's simply more work to finish this implementation.

Luckily for now we only have one patch, which is making master.cf
obsolete and allows to directly specify the path to the main.cf for all
of the Postfix services instead of a configuration directory.

The reason why I implemented that patch by using a sed script instead of
a diff file is that I don't want to rebase a huge patchset of very
redundant changes (mostly renames) on every new Postfix release.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>