services/postfix: Fix whitespace errors in cfgfile
Well, "error" is a bit exaggerated, it's more of a "nitpick" because I
can't stand having unnecessary whitespace to the end of each line.

So this doesn't really fix a real error in the configuration syntax
because it would even be valid with the unnecessary space that's
generated when using lists.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>