services/postfix: Provide nummeric port values
Postfix uses values from /etc/services and also allows for duplicate
definitions of the same service name. The reason is that the service
"name" is actually an address and they only collide of they are defined
multiple times for the same service type.

So for now until we implement service lookup let's keep port numbers
numeric.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>