services/postfix: Allow specifying numeric address
If the value is just an integer, the address resolves to a specific port
on all interfaces. We also do extensive checking on the port number as
well, so people can't accidentally specify things like 893719823 as a
port number.

And yes, a valid port number begins at 0, because 0 means "pick a random
port number".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>