dns-server: Enable dyndns service
This has been in production since a while and I'm using and relying on
it on a daily basis. And so far it works without issues, even better
than most large dynamic DNS providers out there.

Of course, this doesn't include the credentials and I still need to find
a good way to handle secrets via a deployment where the source is
publicly available.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>