services/postfix: Use socket as stdin for services
The Postfix services are structured a bit like inetd, so they _usually_
won't be satisfied with an FD being passed, so pass them as stdin.

Note that we *do* have a service type that accepts FDs ("pass"), which
we need to implement at some later point after figuring out *how*
they're passed.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>