services/postfix: Fix up and finalize users/groups
In my weird mental state, I somehow inverted the checks for the default
value of the user/group options, so that by default no default user gets
created, which is what we obviously don't want.

Also, the postdrop group was missing, so I've added it along with an
option to be able to change it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>