services/postfix: Allow multiple addrs in services
It's more or less in preparation of how Postfix *really* handles its
services, because our way to use Accept = true is wrong and Postfix
*actually* does something more close to what systemd socket activation
actually already does (passing file descriptors to services that pick
them up but have their own handler for accept()).

So for passing the socket FDs we need to have an upper bound for the
file descriptors, which allows for multiple different listening
addresses.

This has to be supported on our side as well, so what we do now is
simply use the "address" option as a singleton default for "addresses",
which is a list of listening addresses.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>