services/postfix: Flesh out default services conf
We only had a rudimentary definition of the services only vaguely
representing the default master.cf from the upstream project.

With this commit we now have a configuration that more closely resembles
the upstream master.cf using a few helper functions that let us specify
the configuration in a compact way similar to the upstream master.cf.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>