services/postfix: Provide a temporary default conf
It's just setting the compatibility level and an empty aliases file in
the ugliest way so that we're better able to test while implementing the
missing features on our side.

Of course I've marked both of them with XXX so we can make sure that we
remove it before this goes into production.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>