modules/epmd: Fix build of patched EPMD
Since a while pkgs.systemd has multiple outputs, so we need to adapt
this. Also we're now running the full configure script of the OTP
distribution in order to make sure that we have all necessary files
prior to building epmd.

Of course this makes the configurePhase a bit longer than we had so far,
but it's more future-proof that way.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>