pkgs/mongooseim: Update to version 2.0.1
This updates MongooseIM and all (test) dependencies to version 2.0.1.

The changelog is quite huge, so instead of quoting it here, this is the
upstream URL of all changes:

https://github.com/esl/MongooseIM/blob/2.0.1/CHANGELOG.md

Unfortunately we still have a lot of work to do in order to fixup the
tests and actually getting the server to work.

But apart from that every dependency should be updated and every patch
should be rebased against the new version.

The dhparams.patch now is rebased against the fast_tls library and it's
also used as a patch for this library instead of directly applying it to
the MongooseIM sources.

I've also dropped the tests.patch from the MongooseIM tests, because we
need to re-do it from scratch as a lot of what the patch addresses has
been refactored.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>