lib: Refactor credentials system
It makes more sense to not expose credential attributes directly but
instead expose them via a "getcred" function which throws an error if
the credentials file is available but the credential hasn't been found.

The reason behind this is that we want to avoid getting the default
values for credentials if we manage to either introduce a typo in the
credentials file or we just forget to include a certain attribute.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>