xmppoke/genreport: Fix up short open tags
Since short_open_tag is a configuration option and we can't rely on it
to be enabled by default (well, it isn't as of NixOS 16.09), let's make
sure we don't have any short open tags in the report generation script.

We now also pass -f to PHP so that we also catch fatal errors if there
should be any in the future.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>