xmppoke: Fix quoting of LUA_PATH/LUA_CPATH
I have no idea why the single quotes got in there, but after stracing
the process, it clearly tries to search for paths beginning/ending with
a single quote, so this doesn't make sense.

However, I'm still wondering why this has worked in the first place
given that we still use LUA 5.1. But anyway, no time now to dig into
this in detail.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>