mongooseim: Add support for PostgreSQL
This implementation abstracts away dependency handling for all backends
rather than just PostgreSQL, but right now the only package that's in
erlangPackages is p1_pgsql (added with this very commit) so while we can
easily add support for other backends in the future we only support
PostgreSQL right now, because that's what the Headcounter network
already uses.

With this, we also no longer need the strip-unneeded-deps.patch, because
we're doing exactly the same but based on which backends are configured.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>