services/mongooseim: Add options for ODBC
Not directly ODBC under the hood because in case of MySQL and
PostgreSQL (the only database types we support at the moment) MongooseIM
uses an ODBC wrapper around p1_pgsql and p1_mysql, so we don't really
need Erlang with ODBC support.

The options are quite straightforward and are only configured whenever
the type has been set to non-null.

We also make sure that we order the MongooseIM service *after* the
configured database management system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>