mongooseim/modules: Rename "mam" to "mam_meta"
We could use mod_mam directly but we'd need a lot of auxiliary modules
in order to have full MAM support and mod_mam_meta takes care of this.

Of course, we'll loose a bit of configurability because of that (well,
except by using extraConfig of course) but we can still add the other
modules later if we desire to do so.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>