tests/mongooseim: Enable MAM module
This uses PostgreSQL as the storage backend and it's also limiting the
max_connections to exactly the pool sizes we expect MongooseIM to use.

The reason for limiting max_connections that way is that we want to make
sure that MongooseIM doesn't accidentally use more connections.

We also use quite strict permissions (well, not as strict as I'd like it
to be, but stricter than allowing everything) to make sure MongooseIM
won't execute SQL statements we don't want it to execute.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>