services/mongooseim: Update default module config
We have few new modules (blocking, csi, pubsub) in MongooseIM version
2.0.1 which we need to define and also set more or less reasonable
defaults.

I say "more or less", because I'm probably going to refine this in the
long term to be actually more transparent on which module is defined or
not without knowing the various gory details about the configuration
system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>