services/mongooseim: Disable mod_amp by default
The security policy[1] from XEP-0079 is not yet implemented as of
MongooseIM 2.0.1[2] and thus this can reveal presence information about
users on the server. While mod_amp is very useful indeed, this point
really is a deal breaker for deploying this in production so let's
disable the module by default (and thus also in our production config).

[1]: http://xmpp.org/extensions/xep-0079.html#security
[2]: https://mongooseim.readthedocs.io/en/2.0.1/developers-guide/mod_amp_developers_guide/

Signed-off-by: aszlig <aszlig@redmoonstudios.org>