services/mongooseim/modules: Document "options"
This should give a better documentation on the format that is allowed
for MongooseIM module options. In the end we might want to implement a
wrapper around mkOption with a custom type to make this a bit more
uniform accross all of the MongooseIM options that allow to directly
pass over Nix expressions to the resulting Erlang configuration file.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>