services/mongooseim: Put real config in configFile
It's a bit silly to reference the config file on a case-basis by either
using the configFile value if the user has set it or by using an
entirely different option if not.

The reason this is silly is because we can't therefore rely on using
configFile from within another module to get the location of the config
file.

So now it's possible to *always* refer to the correct location of the
config file by using config.headcounter.services.mongooseim.configFile.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>