tests/mongooseim: Check correctness of listeners
We really want to make sure that nothing else listens on the host than
we have explicitly specified in the listeners option.

This is a bit more involved and we need to dig around in various Erlang
internals. One example is getting the net_kernel port that is used to
get connections from other distributed nodes.

I haven't found a way to simply get that port from within net_kernel's
internal state, so we use erlang:process_info/1 here to fish out the
port that listens to the connection from the links that net_kernel has.

A very similar method is done afterwards, when we gather a list of ports
and check which of them are listeners so we can compare them against
what we have defined in our config.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>