Add a config module for configuring Erlang hosts
This is so that we can override whatever is set by /etc/hosts inside of
an Erlang VM. We still fall back to using the native lookup method, but
consider an example /etc/hosts entry like this:

192.168.0.1 examplehost

If we now use -sname foo@examplehost, the inet_tcp_dist module is going
to try binding to 0.0.0.0 in some cases.

However, if we want to have a distributed Erlang only using either the
loopback interface or a certain selected interface, we can't simply use
inet_dist_use_interface alone to enforce binding to 127.0.0.1 because in
some cases like -remsh it still tries binding to 0.0.0.0.

Now if we have a way to explicitly state the node IP we can prevent
this, but we still need to set inet_dist_use_interface later to isolate
it for sure.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>