services/mongooseim: Enforce using loopback iface
This sets inet_dist_use_interface and uses the inetrc from our
erlang-inet NixOS module we have introduced in the previous commit.

The enforcement is only in place whenever epmd.addresses only contains
"127.0.0.1", so it's still easily possible to use this for example
within a private network.

We now also have a now nodeIp option in services.mongooseim, so we can
explicitly set the nodeIp to whatever we want, regardless of what we
have in /etc/hosts.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>