lib/erlang: Introduce a dedicated type erlPropList
Using "hclib.erlType hclib.erlPropList" works fine for scalar types, but
if it comes to containers, we get an eval error as soon as we have
defined multiple attributes in different modules.

This new dedicated erlPropList type now merges the top-level attribute
set the same way as types.attrs in <nixpkgs>.

Emphasis is on "merges the top-level attribute", because if we want to
merge deeper nested attributes, they're not properly merged yet.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>