services/mongooseim: Add listener options for HTTP
So far it has been a bit annoying to configure HTTP/cowboy modules:

  options.modules = [
    { tuple = [
        server1
        "/api"
        { atom = "mongoose_api"; }
        { handlers = [
            { atom = "mongoose_api_metrics"; }
            { atom = "mongoose_api_users"; }
          ];
        }
      ];
    }
    { tuple = ["_" "/http-bind" { atom = "mod_bosh"; }]; }
    { tuple = ["_" "/ws-xmpp"   { atom = "mod_websockets"; }]; }
  ];

This now should be come something like this:

  http.modules = [
    { host = server1;
      path = "/api";
      handler = "mongoose_api";
      options.handlers = [
        { atom = "mongoose_api_metrics"; }
        { atom = "mongoose_api_users"; }
      ];
    }
    { path = "/http-bind";
      handler = "mod_bosh";
    }
    { path = "/ws-xmpp";
      handler = "mod_websockets";
    }
  ];

While this doesn't make a lot of difference in terms of the line count
it is less nested and we have type safety on top of it.

Note that I removed the "mongoose_api" handlers with this commit, even
though I used it as an example. The reason for removing it is that it's
deprecated by the upstream project.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>