tests/mongooseim: Add REST API listeners
These are defined by the upstream default configuration, so we need it
on our part as well.

I haven't yet run the tests so I'm not sure whether it improves on the
succeess of the rest_SUITE. We might need to replace one or more
"localhost" strings of the test code there.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>