services/mongooseim: Allow setting auth options
Instead of settingn.authMethod, we now have settings.auth.method, which
is a bit better in terms of namespacing.

And we also have auth.options now, which is a property list that allows
to set a bunch of authentication options.

Unfortunately the auth_opts field is a bit all over the codebase and not
very well documented upstream, so I'm not constraining this yet to only
the options which are _really_ handled by the code.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>