xmpp: Use SCRAM for authentication
It's about time to switch away from plaintext passwords, even though the
passwords are sent via TLS. There is still the chance that we have a
MITM attack on a client that doesn't correctly check the certificate.

And apart from this, we have some rogue CAs that issue certificates for
domains they don't own. WoSign might be one example... *cough* *cough*

Signed-off-by: aszlig <aszlig@redmoonstudios.org>