tests/mongooseim: Incorporate diffs between nodes
If we look at rel/reltool_vars/node[12]_vars.config in the MongooseIM
source tree, we can see that we have a few differences between the first
node and the second. The main difference is that the first node has
mod_offline and mod_amp enabled but the second node hasn't.

So while we want to keep machines separate as much as possible we *do*
want to come more close to the upstream test setup.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>