lib/erlang: Properly escape Erlang strings
If we have an UTF-8 characters within some Nix code, like we introduced
in the previous commit, the resulting Erlang term is a list and it has
the full unicode character number as one of its items.

This in turn causes list_to_binary/1 to fail, because it assumes that
the input consists solely of 8 bit characters.

Now escaping these characters isn't so easy in Nix, because we don't
have a way to convert an integer to a character. So what we're doing
instead is using a generator (lib/erlang/generate.py) which creates an
escaper function in a dedicated Nix expression file.

The source of this escaper is nothing more than a table of the raw
characters transliterated to the hex-encoded equivalents that the
resulting Erlang string will get.

Fortunately the same rules apply if we have an Erlang binary, so that's
covered as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>