pkgs/mongooseim/tests: Remove metrics_api_SUITE
First of all, this requires katt to have its optional dependencies and
second, the tests rely on changing clustering settings, which we don't
support at runtime with our NixOS configuration.

In theory we could keep running the tests of the metrics_api_SUITE
because they're skipped if the optional dependencies of katt are not
available.

However, the init_per_suite/1 function in metrics_api_SUITE.erl looks
like this:

  init_per_suite(Config) ->
      Config1 = dynamic_modules:stop_running(mod_offline, Config),
      Config2 = escalus:init_per_suite(Config1),
      katt_helper:init_per_suite(Config2).

The skip is caused by the last line, but the first line already stopped
mod_offline, which causes a few of the tests that are running after
metrics_api_SUITE to fail.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>