mongooseim/tests: Increase timeout for amp_SUITE
Various variations of notify_deliver_to_malformed_jid_test randomly fail
because the timeout for escalus:wait_for_stanza/2 is too low.

I've also moved a similar increase of timeouts in s2s_SUITE to the main
derivation of the tests, because they aren't test-scenario specific and
thus shouldn't be in tests/mongooseim.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>