pkgs/mongooseim/tests: Remove all skipped suites
All of them are related to either changing the configuration file, which
doesn't work on our side, because the config resides in the Nix store or
they're trying to change the cluster configuration, which also doesn't
work because we have a static configuration on the Nix expression level.

So in order to keep errors that might be caused from the non-atomic side
effects during the setup of these tests to a minimum let's remove them
because in the end we don't care about them anyway and thus there is no
need to even run them.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>