tests/mongooseim: Fix muc-related tests
These are the last remaining tests that have failed so far, so now all
tests should pass.

The reason a few the MUC tests didn't pass is that the test environment
uses HTTP to authenticate rooms and the HTTP server that is used for
authentication is usually assumed to be on localhost.

However, our tests run accross several machines, so the HTTP
authenticator is on the host that runs the tests, which in our case is
"client".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>