tests/headcounter: Refactor for smaller subtests
Instead of generating a bunch of very monolithic per-vhost tests, we now
still have the per-vhosts tests but they're now just a matter of
namespacing the attribute sets and we can throw in whatever test we want
without the need to do weird hacks around make-test.nix.

The way this is implemented is that we wrap make-test.nix in
tests/headcounter/default.nix so that we include common node
configuration like for example the XMPP specific config options for
ultron.

Then we merge every other configuration into the nodes attribute using
imports.

Using this allows us to have one common configuration that is specifying
the test scenario for all other subtests to come. Another thing that's
now possible without refactoring too much is the inclusion of our DNS
configuration to the test scenario.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>