tests/mongooseim: Move CT runner into lib
The test runner generates a nice HTML report which is then picked up by
this script snipped and put up as a Hydra build product.

We're going to write common tests for our vhosts, so we're going to need
the same there as well, hence it makes sense to make it easily
importable.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>