lib/erlang: Allow .freekey in proplists
There are a few occasions, where a key of a proplist isn't an atom but
some other term, like a tuple in "{{s2s_addr, Something}, value}".

In order to support this we can now write something like:

  erlPropList {
    xxx = {
      freekey.tuple = [ 1 2 ];
      value.tuple = [ 3 4 ];
    };
  }

Which will be converted to:

  [{{1, 2}, {3, 4}}]

The reason the xxx is no longer present is because it now acts as some
kind of a dummy key, so that it can still safely be used with other
attributes in the property list without giving a duplicate key error on
the Nix side.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>