lib/erlang: Allow .multi in proplists
Apart from non-atom keys we sometimes have keys that are defined
multiple times in a property list, which now is possible to specify
using Nix expressions:

  erlPropList { xxx.multi = [ 111 222 ]; }

Will result in:

  [{xxx, 111}, {xxx, 222}]

The implementation is a bit more involved, because we now need to splice
in the relevant duplicate items into the list of key/value pairs.

We do this by creating singletons for every operation but mkMulti in
mkTuple, which also means that the guts of erlPropList are now in
erlPropListTerms so that the whole concatenation is done elsewhere and
we can recurse easily on mkTuple.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>